requirements-test.txt
python-escpos>=3.0
pillow>=9.0
